requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
gcloud==0.18.3
googleapis-common-protos==1.6.0
gunicorn==19.9.0
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
jws==0.1.3
MarkupSafe==1.1.1
numpy==1.17.2
oauth2client==4.1.2
pandas==0.25.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycryptodome==3.9.0
Pyrebase4==4.3.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-jwt==2.0.1
pytz==2019.3
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
urllib3==1.25.6
Werkzeug==0.16.0
